refactored xcube.core.mldataset into package
