+++ xcube/core/mldataset/mapped.py
@@ -1,0 +1,55 @@
+# The MIT License (MIT)
+# Copyright (c) 2022 by the xcube development team and contributors
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+# DEALINGS IN THE SOFTWARE.
+
+from typing import Any, Dict, Callable
+
+import xarray as xr
+
+from .abc import MultiLevelDataset
+from .lazy import LazyMultiLevelDataset
+
+
+class MappedMultiLevelDataset(LazyMultiLevelDataset):
+    def __init__(
+            self,
+            ml_dataset: MultiLevelDataset,
+            mapper_function: Callable[[xr.Dataset, ...], xr.Dataset],
+            ds_id: str = None,
+            mapper_params: Dict[str, Any] = None
+    ):
+        """
+
+        """
+        super().__init__(ds_id=ds_id,
+                         parameters=mapper_params)
+        self._ml_dataset = ml_dataset
+        self._mapper_function = mapper_function
+
+    def _get_num_levels_lazily(self) -> int:
+        return self._ml_dataset.num_levels
+
+    def _get_dataset_lazily(self, index: int,
+                            mapper_params: Dict[str, Any]) -> xr.Dataset:
+        return self._mapper_function(self._ml_dataset.get_dataset(index),
+                                     **mapper_params)
+
+    def close(self):
+        self._ml_dataset.close()
